+++ src/Map/MapDraw.cpp
@@ -97,4 +97,4 @@ TMapDrawWall::TMapDrawWall(const char* name)
 {
 }
 
-void TMap::draw(unsigned long, JDrama::TGraphics*) const { }
+void TMap::draw(u32, JDrama::TGraphics*) const { }
